tests: Treat each file as a separate crate
